clean hdoc secure

+++ commun/srcs/utils/clean_hdoc.c
@@ -6,7 +6,7 @@
 /*   By: fmauguin <fmauguin@student.42.fr >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/30 13:48:24 by fmauguin          #+#    #+#             */
-/*   Updated: 2022/06/30 14:49:55 by fmauguin         ###   ########.fr       */
+/*   Updated: 2022/07/02 00:25:29 by fmauguin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,10 @@
 
 static int	unlink_hdoc(char *name)
 {
+	if (!name)
+		return (0);
+	if (access(name, F_OK) == -1)
+		return (0);
 	if (unlink(name) == -1)
 		return (1);
 	return (0);
@@ -29,6 +33,8 @@ int	del_hdoc_token(t_tokens *tokens)
 	size_t	i;
 
 	i = 0;
+	if (!tokens || !tokens->tokens)
+		return (0);
 	while (i < tokens->size)
 	{
 		if (i > 0 && tokens->tokens[i - 1].type == IO_HDOC)
@@ -47,6 +53,8 @@ int	del_hdoc_parse(t_parse **parse)
 	size_t	j;
 
 	i = 0;
+	if (!parse)
+		return (0);
 	while (parse[i])
 	{
 		if (parse[i]->type == CMD && parse[i]->cmd && parse[i]->cmd->redirect)
